Added a feature spec for editing the events for a competition. Turns out this was broken, so I added the missing "=" character, which fixes #953.
